requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
